改长了变量存储空间, 后台最多支持15个人

+++ UnderCoverIOS/UnderCoverIOS/Model/ViewGuessViewController.h
@@ -11,8 +11,8 @@
 @interface ViewGuessViewController : UCIBaseViewController{
     int PeopleCount;
     int SonCount;
-    int fatherSet;
-    int sonSet;
+    long long fatherSet;
+    long long sonSet;
     int MaxPeopleCount;
     int MaxSonCount;
     int curPeople;
